requirements.txt
fastapi
uvicorn
sentence-transformers
chromadb
pdfminer.six
python-docx
aiofiles
python-multipart
